AL6-Player1/2-AudioFiles, AL6-Player2/2-PlaylistSrlz.

+++ Player/PlayingItem.cs
@@ -8,7 +8,7 @@ namespace MyPlayer
 {
     public class PlayingItem : IComparable<PlayingItem>
     {
-        public int Duration;
+        public double Duration;
         public string Name;
 
         public int CompareTo(PlayingItem comparePart)
